scripts/self-signed-cert-utility/requirements.txt
boto3
pyopenssl
validators
